# api/v1/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for DixitMe - Online Dixit Card Game\nThis API provides endpoints for managing players, games, and real-time gameplay through WebSocket connections.",
        "title": "DixitMe API",
        "contact": {
            "name": "DixitMe API Support",
            "email": "support@dixitme.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/history": {
            "get": {
                "description": "Get chat message history for a specific game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "game_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of messages to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of messages to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChatHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/chat/send": {
            "post": {
                "description": "Send a chat message to a specific game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Send chat message",
                "parameters": [
                    {
                        "description": "Chat message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SendChatMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/chat/stats": {
            "get": {
                "description": "Get statistics about chat usage across all games",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChatStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ChatHistoryResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatMessage"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handlers.ChatStatsResponse": {
            "type": "object",
            "properties": {
                "active_games": {
                    "type": "integer"
                },
                "messages_this_week": {
                    "type": "integer"
                },
                "messages_today": {
                    "type": "integer"
                },
                "total_messages": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SendChatMessageRequest": {
            "type": "object",
            "required": [
                "game_id",
                "message"
            ],
            "properties": {
                "game_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "message_type": {
                    "description": "chat, system, emote",
                    "type": "string"
                }
            }
        },
...
        "models.AuthType": {
            "type": "string",
            "enum": [
                "guest",
                "password",
                "google"
            ],
            "x-enum-varnames": [
                "AuthTypeGuest",
                "AuthTypePassword",
                "AuthTypeGoogle"
            ]
        },
...
        "models.CardSubmission": {
            "type": "object",
            "properties": {
                "card_id": {
                    "description": "ID of the card from the deck",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                },
                "round": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GameRound"
                        }
                    ]
                },
                "round_id": {
                    "type": "string"
                }
            }
        },
...
        "models.ChatMessage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "game": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Game"
                        }
                    ]
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_visible": {
                    "description": "For moderation",
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "message_type": {
                    "description": "chat, system, emote",
                    "type": "string"
                },
                "phase": {
                    "description": "lobby, storytelling, submitting, voting, scoring",
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
        "models.Game": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "max_rounds": {
                    "description": "3 players * 2 rounds each",
                    "type": "integer"
                },
                "players": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GamePlayer"
                    }
                },
                "room_code": {
                    "type": "string"
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GameRound"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.GameStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.GamePlayer": {
            "type": "object",
            "properties": {
                "game": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Game"
                        }
                    ]
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                },
                "position": {
                    "description": "Turn order",
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "models.GameRound": {
            "type": "object",
            "properties": {
                "clue": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "game": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Game"
                        }
                    ]
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "round_number": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                },
                "storyteller": {
                    "$ref": "#/definitions/models.Player"
                },
                "storyteller_card": {
                    "description": "Card ID chosen by storyteller",
                    "type": "integer"
                },
                "storyteller_id": {
                    "type": "string"
                },
                "submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CardSubmission"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vote"
                    }
                }
            }
        },
        "models.GameStatus": {
            "type": "string",
            "enum": [
                "waiting",
                "in_progress",
                "completed",
                "abandoned"
            ],
            "x-enum-varnames": [
                "GameStatusWaiting",
                "GameStatusInProgress",
                "GameStatusCompleted",
                "GameStatusAbandoned"
            ]
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "$ref": "#/definitions/models.AuthType"
                },
                "bot_level": {
                    "description": "easy, medium, hard",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.PlayerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "NULL for guest players",
                    "type": "string"
                }
            }
        },
        "models.PlayerType": {
            "type": "string",
            "enum": [
                "human",
                "bot"
            ],
            "x-enum-varnames": [
                "PlayerTypeHuman",
                "PlayerTypeBot"
            ]
        },
        "models.RoundStatus": {
            "type": "string",
            "enum": [
                "storytelling",
                "submitting",
                "voting",
                "scoring",
                "completed"
            ],
            "x-enum-varnames": [
                "RoundStatusStorytelling",
                "RoundStatusSubmitting",
                "RoundStatusVoting",
                "RoundStatusScoring",
                "RoundStatusCompleted"
            ]
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "$ref": "#/definitions/models.AuthType"
                },
                "avatar": {
                    "description": "Profile picture URL",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Vote": {
            "type": "object",
            "properties": {
                "card_id": {
                    "description": "ID of the card they voted for",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                },
                "round": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GameRound"
                        }
                    ]
                },
                "round_id": {
                    "type": "string"
                }
            }
        }
    }
}